Rearrange Species Lists
Commit is a little weird with what it shows as changes, but it was just rearranging stuff
