+++ src/pit_randomizer.c
@@ -1297,6 +1297,11 @@ static const u16 sAvatarSpecies[] =
     // SPECIES_EGG       ,
 };
 
+u32 GetMaxNumberOfSpeciesAvatars(void)
+{
+    return AVATAR_SPECIES_COUNT;
+}
+
 u16 AccessAvatarSpeciesArrayIndex(u16 index)
 {
     return sAvatarSpecies[index % AVATAR_SPECIES_COUNT];
@@ -1312,1281 +1317,6 @@ u16 GetIndexOfSpeciesInAvatarSpeciesArray(u16 species)
         }
     }
     return 0;
-}
-
-
-//
-//		Random Pokemon
-//
-
-//*********************** //tx_randomizer_and_challenges
-#define RANDOM_SPECIES_COUNT_LEGENDARY ARRAY_COUNT(sRandomSpeciesLegendary)
-static const u16 sRandomSpeciesLegendary[] =
-{
-    //SPECIES_NONE                    ,
-    SPECIES_BULBASAUR,
-    SPECIES_IVYSAUR,
-    SPECIES_VENUSAUR,
-    SPECIES_CHARMANDER,
-    SPECIES_CHARMELEON,
-    SPECIES_CHARIZARD,
-    SPECIES_SQUIRTLE,
-    SPECIES_WARTORTLE,
-    SPECIES_BLASTOISE,
-    SPECIES_CATERPIE,
-    SPECIES_METAPOD,
-    SPECIES_BUTTERFREE,
-    SPECIES_WEEDLE,
-    SPECIES_KAKUNA,
-    SPECIES_BEEDRILL,
-    SPECIES_PIDGEY,
-    SPECIES_PIDGEOTTO,
-    SPECIES_PIDGEOT,
-    SPECIES_RATTATA,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_RATTATA_ALOLAN,
-#endif
-    SPECIES_RATICATE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_RATICATE_ALOLAN,
-#endif
-    SPECIES_SPEAROW,
-    SPECIES_FEAROW,
-    SPECIES_EKANS,
-    SPECIES_ARBOK,
-    SPECIES_PIKACHU,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_PIKACHU_HOENN_CAP,
-#endif
-    SPECIES_RAICHU,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_RAICHU_ALOLAN,
-#endif
-    SPECIES_SANDSHREW,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SANDSHREW_ALOLAN,
-#endif
-    SPECIES_SANDSLASH,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SANDSLASH_ALOLAN,
-#endif
-    SPECIES_NIDORAN_F,
-    SPECIES_NIDORINA,
-    SPECIES_NIDOQUEEN,
-    SPECIES_NIDORAN_M,
-    SPECIES_NIDORINO,
-    SPECIES_NIDOKING,
-    SPECIES_CLEFAIRY,
-    SPECIES_CLEFABLE,
-    SPECIES_VULPIX,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_VULPIX_ALOLAN,
-#endif
-    SPECIES_NINETALES,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_NINETALES_ALOLAN,
-#endif
-    SPECIES_JIGGLYPUFF,
-    SPECIES_WIGGLYTUFF,
-    SPECIES_ZUBAT,
-    SPECIES_GOLBAT,
-    SPECIES_ODDISH,
-    SPECIES_GLOOM,
-    SPECIES_VILEPLUME,
-    SPECIES_PARAS,
-    SPECIES_PARASECT,
-    SPECIES_VENONAT,
-    SPECIES_VENOMOTH,
-    SPECIES_DIGLETT,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_DIGLETT_ALOLAN,
-#endif
-    SPECIES_DUGTRIO,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_DUGTRIO_ALOLAN,
-#endif
-    SPECIES_MEOWTH,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_MEOWTH_ALOLAN,
-    SPECIES_MEOWTH_GALARIAN,
-#endif
-    SPECIES_PERSIAN,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_PERSIAN_ALOLAN,
-#endif
-    SPECIES_PSYDUCK,
-    SPECIES_GOLDUCK,
-    SPECIES_MANKEY,
-    SPECIES_PRIMEAPE,
-    SPECIES_GROWLITHE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_GROWLITHE_HISUIAN,
-#endif
-    SPECIES_ARCANINE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ARCANINE_HISUIAN,
-#endif
-    SPECIES_POLIWAG,
-    SPECIES_POLIWHIRL,
-    SPECIES_POLIWRATH,
-    SPECIES_ABRA,
-    SPECIES_KADABRA,
-    SPECIES_ALAKAZAM,
-    SPECIES_MACHOP,
-    SPECIES_MACHOKE,
-    SPECIES_MACHAMP,
-    SPECIES_BELLSPROUT,
-    SPECIES_WEEPINBELL,
-    SPECIES_VICTREEBEL,
-    SPECIES_TENTACOOL,
-    SPECIES_TENTACRUEL,
-    SPECIES_GEODUDE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_GEODUDE_ALOLAN,
-#endif
-    SPECIES_GRAVELER,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_GRAVELER_ALOLAN,
-#endif
-    SPECIES_GOLEM,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_GOLEM_ALOLAN,
-#endif
-    SPECIES_PONYTA,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_PONYTA_GALARIAN,
-#endif
-    SPECIES_RAPIDASH,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_RAPIDASH_GALARIAN,
-#endif
-    SPECIES_SLOWPOKE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SLOWPOKE_GALARIAN,
-#endif
-    SPECIES_SLOWBRO,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SLOWBRO_GALARIAN,
-#endif
-    SPECIES_MAGNEMITE,
-    SPECIES_MAGNETON,
-    SPECIES_FARFETCHD,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_FARFETCHD_GALARIAN,
-#endif
-    SPECIES_DODUO,
-    SPECIES_DODRIO,
-    SPECIES_SEEL,
-    SPECIES_DEWGONG,
-    SPECIES_GRIMER,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_GRIMER_ALOLAN,
-#endif
-    SPECIES_MUK,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_MUK_ALOLAN,
-#endif
-    SPECIES_SHELLDER,
-    SPECIES_CLOYSTER,
-    SPECIES_GASTLY,
-    SPECIES_HAUNTER,
-    SPECIES_GENGAR,
-    SPECIES_ONIX,
-    SPECIES_DROWZEE,
-    SPECIES_HYPNO,
-    SPECIES_KRABBY,
-    SPECIES_KINGLER,
-    SPECIES_VOLTORB,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_VOLTORB_HISUIAN,
-#endif
-    SPECIES_ELECTRODE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ELECTRODE_HISUIAN,
-#endif
-    SPECIES_EXEGGCUTE,
-    SPECIES_EXEGGUTOR,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_EXEGGUTOR_ALOLAN,
-#endif
-    SPECIES_CUBONE,
-    SPECIES_MAROWAK,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_MAROWAK_ALOLAN,
-#endif
-    SPECIES_HITMONLEE,
-    SPECIES_HITMONCHAN,
-    SPECIES_LICKITUNG,
-    SPECIES_KOFFING,
-    SPECIES_WEEZING,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_WEEZING_GALARIAN,
-#endif
-    SPECIES_RHYHORN,
-    SPECIES_RHYDON,
-    SPECIES_CHANSEY,
-    SPECIES_TANGELA,
-    SPECIES_KANGASKHAN,
-    SPECIES_HORSEA,
-    SPECIES_SEADRA,
-    SPECIES_GOLDEEN,
-    SPECIES_SEAKING,
-    SPECIES_STARYU,
-    SPECIES_STARMIE,
-    SPECIES_MR_MIME,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_MR_MIME_GALARIAN,
-#endif
-    SPECIES_SCYTHER,
-    SPECIES_JYNX,
-    SPECIES_ELECTABUZZ,
-    SPECIES_MAGMAR,
-    SPECIES_PINSIR,
-    SPECIES_TAUROS,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_TAUROS_PALDEAN_COMBAT_BREED,
-    SPECIES_TAUROS_PALDEAN_BLAZE_BREED,
-    SPECIES_TAUROS_PALDEAN_AQUA_BREED,
-#endif
-    SPECIES_MAGIKARP,
-    SPECIES_GYARADOS,
-    SPECIES_LAPRAS,
-    SPECIES_DITTO,
-    SPECIES_EEVEE,
-    SPECIES_VAPOREON,
-    SPECIES_JOLTEON,
-    SPECIES_FLAREON,
-    SPECIES_PORYGON,
-    SPECIES_OMANYTE,
-    SPECIES_OMASTAR,
-    SPECIES_KABUTO,
-    SPECIES_KABUTOPS,
-    SPECIES_AERODACTYL,
-    SPECIES_SNORLAX,
-    SPECIES_ARTICUNO,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ARTICUNO_GALARIAN,
-#endif
-    SPECIES_ZAPDOS,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ZAPDOS_GALARIAN,
-#endif
-    SPECIES_MOLTRES,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_MOLTRES_GALARIAN,
-#endif
-    SPECIES_DRATINI,
-    SPECIES_DRAGONAIR,
-    SPECIES_DRAGONITE,
-    SPECIES_MEWTWO,
-    SPECIES_MEW,
-    SPECIES_CHIKORITA,
-    SPECIES_BAYLEEF,
-    SPECIES_MEGANIUM,
-    SPECIES_CYNDAQUIL,
-    SPECIES_QUILAVA,
-    SPECIES_TYPHLOSION,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_TYPHLOSION_HISUIAN,
-#endif
-    SPECIES_TOTODILE,
-    SPECIES_CROCONAW,
-    SPECIES_FERALIGATR,
-    SPECIES_SENTRET,
-    SPECIES_FURRET,
-    SPECIES_HOOTHOOT,
-    SPECIES_NOCTOWL,
-    SPECIES_LEDYBA,
-    SPECIES_LEDIAN,
-    SPECIES_SPINARAK,
-    SPECIES_ARIADOS,
-    SPECIES_CROBAT,
-    SPECIES_CHINCHOU,
-    SPECIES_LANTURN,
-    SPECIES_PICHU,
-#ifndef PIT_GEN_3_MODE
-    SPECIES_PICHU_SPIKY_EARED,
-#endif
-    SPECIES_CLEFFA,
-    SPECIES_IGGLYBUFF,
-    SPECIES_TOGEPI,
-    SPECIES_TOGETIC,
-    SPECIES_NATU,
-    SPECIES_XATU,
-    SPECIES_MAREEP,
-    SPECIES_FLAAFFY,
-    SPECIES_AMPHAROS,
-    SPECIES_BELLOSSOM,
-    SPECIES_MARILL,
-    SPECIES_AZUMARILL,
-    SPECIES_SUDOWOODO,
-    SPECIES_POLITOED,
-    SPECIES_HOPPIP,
-    SPECIES_SKIPLOOM,
-    SPECIES_JUMPLUFF,
-    SPECIES_AIPOM,
-    SPECIES_SUNKERN,
-    SPECIES_SUNFLORA,
-    SPECIES_YANMA,
-    SPECIES_WOOPER,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_WOOPER_PALDEAN,
-#endif
-    SPECIES_QUAGSIRE,
-    SPECIES_ESPEON,
-    SPECIES_UMBREON,
-    SPECIES_MURKROW,
-    SPECIES_SLOWKING,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SLOWKING_GALARIAN,
-#endif
-    SPECIES_MISDREAVUS,
-    SPECIES_UNOWN,
-    SPECIES_WOBBUFFET,
-    SPECIES_GIRAFARIG,
-    SPECIES_PINECO,
-    SPECIES_FORRETRESS,
-    SPECIES_DUNSPARCE,
-    SPECIES_GLIGAR,
-    SPECIES_STEELIX,
-    SPECIES_SNUBBULL,
-    SPECIES_GRANBULL,
-    SPECIES_QWILFISH,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_QWILFISH_HISUIAN,
-#endif
-    SPECIES_SCIZOR,
-    SPECIES_SHUCKLE,
-    SPECIES_HERACROSS,
-    SPECIES_SNEASEL,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SNEASEL_HISUIAN,
-#endif
-    SPECIES_TEDDIURSA,
-    SPECIES_URSARING,
-    SPECIES_SLUGMA,
-    SPECIES_MAGCARGO,
-    SPECIES_SWINUB,
-    SPECIES_PILOSWINE,
-    SPECIES_CORSOLA,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_CORSOLA_GALARIAN,
-#endif
-    SPECIES_REMORAID,
-    SPECIES_OCTILLERY,
-    SPECIES_DELIBIRD,
-    SPECIES_MANTINE,
-    SPECIES_SKARMORY,
-    SPECIES_HOUNDOUR,
-    SPECIES_HOUNDOOM,
-    SPECIES_KINGDRA,
-    SPECIES_PHANPY,
-    SPECIES_DONPHAN,
-    SPECIES_PORYGON2,
-    SPECIES_STANTLER,
-    SPECIES_SMEARGLE,
-    SPECIES_TYROGUE,
-    SPECIES_HITMONTOP,
-    SPECIES_SMOOCHUM,
-    SPECIES_ELEKID,
-    SPECIES_MAGBY,
-    SPECIES_MILTANK,
-    SPECIES_BLISSEY,
-    SPECIES_RAIKOU,
-    SPECIES_ENTEI,
-    SPECIES_SUICUNE,
-    SPECIES_LARVITAR,
-    SPECIES_PUPITAR,
-    SPECIES_TYRANITAR,
-    SPECIES_LUGIA,
-    SPECIES_HO_OH,
-    SPECIES_CELEBI,
-    SPECIES_TREECKO,
-    SPECIES_GROVYLE,
-    SPECIES_SCEPTILE,
-    SPECIES_TORCHIC,
-    SPECIES_COMBUSKEN,
-    SPECIES_BLAZIKEN,
-    SPECIES_MUDKIP,
-    SPECIES_MARSHTOMP,
-    SPECIES_SWAMPERT,
-    SPECIES_POOCHYENA,
-    SPECIES_MIGHTYENA,
-    SPECIES_ZIGZAGOON,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ZIGZAGOON_GALARIAN,
-#endif
-    SPECIES_LINOONE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_LINOONE_GALARIAN,
-#endif
-    SPECIES_WURMPLE,
-    SPECIES_SILCOON,
-    SPECIES_BEAUTIFLY,
-    SPECIES_CASCOON,
-    SPECIES_DUSTOX,
-    SPECIES_LOTAD,
-    SPECIES_LOMBRE,
-    SPECIES_LUDICOLO,
-    SPECIES_SEEDOT,
-    SPECIES_NUZLEAF,
-    SPECIES_SHIFTRY,
-    SPECIES_TAILLOW,
-    SPECIES_SWELLOW,
-    SPECIES_WINGULL,
-    SPECIES_PELIPPER,
-    SPECIES_RALTS,
-    SPECIES_KIRLIA,
-    SPECIES_GARDEVOIR,
-    SPECIES_SURSKIT,
-    SPECIES_MASQUERAIN,
-    SPECIES_SHROOMISH,
-    SPECIES_BRELOOM,
-    SPECIES_SLAKOTH,
-    SPECIES_VIGOROTH,
-    SPECIES_SLAKING,
-    SPECIES_NINCADA,
-    SPECIES_NINJASK,
-    //SPECIES_SHEDINJA,
-    SPECIES_WHISMUR,
-    SPECIES_LOUDRED,
-    SPECIES_EXPLOUD,
-    SPECIES_MAKUHITA,
-    SPECIES_HARIYAMA,
-    SPECIES_AZURILL,
-    SPECIES_NOSEPASS,
-    SPECIES_SKITTY,
-    SPECIES_DELCATTY,
-    SPECIES_SABLEYE,
-    SPECIES_MAWILE,
-    SPECIES_ARON,
-    SPECIES_LAIRON,
-    SPECIES_AGGRON,
-    SPECIES_MEDITITE,
-    SPECIES_MEDICHAM,
-    SPECIES_ELECTRIKE,
-    SPECIES_MANECTRIC,
-    SPECIES_PLUSLE,
-    SPECIES_MINUN,
-    SPECIES_VOLBEAT,
-    SPECIES_ILLUMISE,
-    SPECIES_ROSELIA,
-    SPECIES_GULPIN,
-    SPECIES_SWALOT,
-    SPECIES_CARVANHA,
-    SPECIES_SHARPEDO,
-    SPECIES_WAILMER,
-    SPECIES_WAILORD,
-    SPECIES_NUMEL,
-    SPECIES_CAMERUPT,
-    SPECIES_TORKOAL,
-    SPECIES_SPOINK,
-    SPECIES_GRUMPIG,
-    SPECIES_SPINDA,
-    SPECIES_TRAPINCH,
-    SPECIES_VIBRAVA,
-    SPECIES_FLYGON,
-    SPECIES_CACNEA,
-    SPECIES_CACTURNE,
-    SPECIES_SWABLU,
-    SPECIES_ALTARIA,
-    SPECIES_ZANGOOSE,
-    SPECIES_SEVIPER,
-    SPECIES_LUNATONE,
-    SPECIES_SOLROCK,
-    SPECIES_BARBOACH,
-    SPECIES_WHISCASH,
-    SPECIES_CORPHISH,
-    SPECIES_CRAWDAUNT,
-    SPECIES_BALTOY,
-    SPECIES_CLAYDOL,
-    SPECIES_LILEEP,
-    SPECIES_CRADILY,
-    SPECIES_ANORITH,
-    SPECIES_ARMALDO,
-    SPECIES_FEEBAS,
-    SPECIES_MILOTIC,
-    SPECIES_CASTFORM_NORMAL,
-    SPECIES_KECLEON,
-    SPECIES_SHUPPET,
-    SPECIES_BANETTE,
-    SPECIES_DUSKULL,
-    SPECIES_DUSCLOPS,
-    SPECIES_TROPIUS,
-    SPECIES_CHIMECHO,
-    SPECIES_ABSOL,
-    SPECIES_WYNAUT,
-    SPECIES_SNORUNT,
-    SPECIES_GLALIE,
-    SPECIES_SPHEAL,
-    SPECIES_SEALEO,
-    SPECIES_WALREIN,
-    SPECIES_CLAMPERL,
-    SPECIES_HUNTAIL,
-    SPECIES_GOREBYSS,
-    SPECIES_RELICANTH,
-    SPECIES_LUVDISC,
-    SPECIES_BAGON,
-    SPECIES_SHELGON,
-    SPECIES_SALAMENCE,
-    SPECIES_BELDUM,
-    SPECIES_METANG,
-    SPECIES_METAGROSS,
-    SPECIES_REGIROCK,
-    SPECIES_REGICE,
-    SPECIES_REGISTEEL,
-    SPECIES_LATIAS,
-    SPECIES_LATIOS,
-    SPECIES_KYOGRE,
-    SPECIES_GROUDON,
-    SPECIES_RAYQUAZA,
-    SPECIES_JIRACHI,
-    SPECIES_DEOXYS_NORMAL,
-#ifndef PIT_GEN_3_MODE
-    SPECIES_TURTWIG,
-    SPECIES_GROTLE,
-    SPECIES_TORTERRA,
-    SPECIES_CHIMCHAR,
-    SPECIES_MONFERNO,
-    SPECIES_INFERNAPE,
-    SPECIES_PIPLUP,
-    SPECIES_PRINPLUP,
-    SPECIES_EMPOLEON,
-    SPECIES_STARLY,
-    SPECIES_STARAVIA,
-    SPECIES_STARAPTOR,
-    SPECIES_BIDOOF,
-    SPECIES_BIBAREL,
-    SPECIES_KRICKETOT,
-    SPECIES_KRICKETUNE,
-    SPECIES_SHINX,
-    SPECIES_LUXIO,
-    SPECIES_LUXRAY,
-    SPECIES_BUDEW,
-    SPECIES_ROSERADE,
-    SPECIES_CRANIDOS,
-    SPECIES_RAMPARDOS,
-    SPECIES_SHIELDON,
-    SPECIES_BASTIODON,
-    SPECIES_BURMY_PLANT_CLOAK,
-    SPECIES_BURMY_SANDY_CLOAK,
-    SPECIES_BURMY_TRASH_CLOAK,
-    SPECIES_WORMADAM_PLANT_CLOAK,
-    SPECIES_WORMADAM_SANDY_CLOAK,
-    SPECIES_WORMADAM_TRASH_CLOAK,
-    SPECIES_MOTHIM_PLANT_CLOAK,
-    SPECIES_COMBEE,
-    SPECIES_VESPIQUEN,
-    SPECIES_PACHIRISU,
-    SPECIES_BUIZEL,
-    SPECIES_FLOATZEL,
-    SPECIES_CHERUBI,
-    SPECIES_CHERRIM_OVERCAST,
-    SPECIES_CHERRIM_SUNSHINE,
-    SPECIES_SHELLOS_WEST_SEA,
-    SPECIES_GASTRODON_WEST_SEA,
-    SPECIES_AMBIPOM,
-    SPECIES_DRIFLOON,
-    SPECIES_DRIFBLIM,
-    SPECIES_BUNEARY,
-    SPECIES_LOPUNNY,
-    SPECIES_MISMAGIUS,
-    SPECIES_HONCHKROW,
-    SPECIES_GLAMEOW,
-    SPECIES_PURUGLY,
-    SPECIES_CHINGLING,
-    SPECIES_STUNKY,
-    SPECIES_SKUNTANK,
-    SPECIES_BRONZOR,
-    SPECIES_BRONZONG,
-    SPECIES_BONSLY,
-    SPECIES_MIME_JR,
-    SPECIES_HAPPINY,
-    SPECIES_CHATOT,
-    SPECIES_SPIRITOMB,
-    SPECIES_GIBLE,
-    SPECIES_GABITE,
-    SPECIES_GARCHOMP,
-    SPECIES_MUNCHLAX,
-    SPECIES_RIOLU,
-    SPECIES_LUCARIO,
-    SPECIES_HIPPOPOTAS,
-    SPECIES_HIPPOWDON,
-    SPECIES_SKORUPI,
-    SPECIES_DRAPION,
-    SPECIES_CROAGUNK,
-    SPECIES_TOXICROAK,
-    SPECIES_CARNIVINE,
-    SPECIES_FINNEON,
-    SPECIES_LUMINEON,
-    SPECIES_MANTYKE,
-    SPECIES_SNOVER,
-    SPECIES_ABOMASNOW,
-    SPECIES_WEAVILE,
-    SPECIES_MAGNEZONE,
-    SPECIES_LICKILICKY,
-    SPECIES_RHYPERIOR,
-    SPECIES_TANGROWTH,
-    SPECIES_ELECTIVIRE,
-    SPECIES_MAGMORTAR,
-    SPECIES_TOGEKISS,
-    SPECIES_YANMEGA,
-    SPECIES_LEAFEON,
-    SPECIES_GLACEON,
-    SPECIES_GLISCOR,
-    SPECIES_MAMOSWINE,
-    SPECIES_PORYGON_Z,
-    SPECIES_GALLADE,
-    SPECIES_PROBOPASS,
-    SPECIES_DUSKNOIR,
-    SPECIES_FROSLASS,
-    SPECIES_ROTOM,
-    SPECIES_UXIE,
-    SPECIES_MESPRIT,
-    SPECIES_AZELF,
-    SPECIES_DIALGA,
-    SPECIES_PALKIA,
-    SPECIES_HEATRAN,
-    SPECIES_REGIGIGAS,
-    SPECIES_GIRATINA_ALTERED,
-    SPECIES_CRESSELIA,
-    SPECIES_PHIONE,
-    SPECIES_MANAPHY,
-    SPECIES_DARKRAI,
-    SPECIES_SHAYMIN_LAND,
-    SPECIES_ARCEUS_NORMAL,
-    SPECIES_VICTINI,
-    SPECIES_SNIVY,
-    SPECIES_SERVINE,
-    SPECIES_SERPERIOR,
-    SPECIES_TEPIG,
-    SPECIES_PIGNITE,
-    SPECIES_EMBOAR,
-    SPECIES_OSHAWOTT,
-    SPECIES_DEWOTT,
-    SPECIES_SAMUROTT,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_SAMUROTT_HISUIAN,
-#endif
-    SPECIES_PATRAT,
-    SPECIES_WATCHOG,
-    SPECIES_LILLIPUP,
-    SPECIES_HERDIER,
-    SPECIES_STOUTLAND,
-    SPECIES_PURRLOIN,
-    SPECIES_LIEPARD,
-    SPECIES_PANSAGE,
-    SPECIES_SIMISAGE,
-    SPECIES_PANSEAR,
-    SPECIES_SIMISEAR,
-    SPECIES_PANPOUR,
-    SPECIES_SIMIPOUR,
-    SPECIES_MUNNA,
-    SPECIES_MUSHARNA,
-    SPECIES_PIDOVE,
-    SPECIES_TRANQUILL,
-    SPECIES_UNFEZANT,
-    SPECIES_BLITZLE,
-    SPECIES_ZEBSTRIKA,
-    SPECIES_ROGGENROLA,
-    SPECIES_BOLDORE,
-    SPECIES_GIGALITH,
-    SPECIES_WOOBAT,
-    SPECIES_SWOOBAT,
-    SPECIES_DRILBUR,
-    SPECIES_EXCADRILL,
-    SPECIES_AUDINO,
-    SPECIES_TIMBURR,
-    SPECIES_GURDURR,
-    SPECIES_CONKELDURR,
-    SPECIES_TYMPOLE,
-    SPECIES_PALPITOAD,
-    SPECIES_SEISMITOAD,
-    SPECIES_THROH,
-    SPECIES_SAWK,
-    SPECIES_SEWADDLE,
-    SPECIES_SWADLOON,
-    SPECIES_LEAVANNY,
-    SPECIES_VENIPEDE,
-    SPECIES_WHIRLIPEDE,
-    SPECIES_SCOLIPEDE,
-    SPECIES_COTTONEE,
-    SPECIES_WHIMSICOTT,
-    SPECIES_PETILIL,
-    SPECIES_LILLIGANT,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_LILLIGANT_HISUIAN,
-#endif
-    SPECIES_BASCULIN_RED_STRIPED,
-    SPECIES_SANDILE,
-    SPECIES_KROKOROK,
-    SPECIES_KROOKODILE,
-    SPECIES_DARUMAKA,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_DARUMAKA_GALARIAN,
-#endif
-    SPECIES_DARMANITAN_STANDARD_MODE,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_DARMANITAN_GALARIAN_STANDARD_MODE,
-#endif
-    SPECIES_MARACTUS,
-    SPECIES_DWEBBLE,
-    SPECIES_CRUSTLE,
-    SPECIES_SCRAGGY,
-    SPECIES_SCRAFTY,
-    SPECIES_SIGILYPH,
-    SPECIES_YAMASK,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_YAMASK_GALARIAN,
-#endif
-    SPECIES_COFAGRIGUS,
-    SPECIES_TIRTOUGA,
-    SPECIES_CARRACOSTA,
-    SPECIES_ARCHEN,
-    SPECIES_ARCHEOPS,
-    SPECIES_TRUBBISH,
-    SPECIES_GARBODOR,
-    SPECIES_ZORUA,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ZORUA_HISUIAN,
-#endif
-    SPECIES_ZOROARK,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_ZOROARK_HISUIAN,
-#endif
-    SPECIES_MINCCINO,
-    SPECIES_CINCCINO,
-    SPECIES_GOTHITA,
-    SPECIES_GOTHORITA,
-    SPECIES_GOTHITELLE,
-    SPECIES_SOLOSIS,
-    SPECIES_DUOSION,
-    SPECIES_REUNICLUS,
-    SPECIES_DUCKLETT,
-    SPECIES_SWANNA,
-    SPECIES_VANILLITE,
-    SPECIES_VANILLISH,
-    SPECIES_VANILLUXE,
-    SPECIES_DEERLING_SPRING,
-    SPECIES_DEERLING_SUMMER,
-    SPECIES_DEERLING_AUTUMN,
-    SPECIES_DEERLING_WINTER,
-    SPECIES_SAWSBUCK_SPRING,
-    SPECIES_SAWSBUCK_SUMMER,
-    SPECIES_SAWSBUCK_AUTUMN,
-    SPECIES_SAWSBUCK_WINTER,
-    SPECIES_EMOLGA,
-    SPECIES_KARRABLAST,
-    SPECIES_ESCAVALIER,
-    SPECIES_FOONGUS,
-    SPECIES_AMOONGUSS,
-    SPECIES_FRILLISH,
-    SPECIES_JELLICENT,
-    SPECIES_ALOMOMOLA,
-    SPECIES_JOLTIK,
-    SPECIES_GALVANTULA,
-    SPECIES_FERROSEED,
-    SPECIES_FERROTHORN,
-    SPECIES_KLINK,
-    SPECIES_KLANG,
-    SPECIES_KLINKLANG,
-    SPECIES_TYNAMO,
-    SPECIES_EELEKTRIK,
-    SPECIES_EELEKTROSS,
-    SPECIES_ELGYEM,
-    SPECIES_BEHEEYEM,
-    SPECIES_LITWICK,
-    SPECIES_LAMPENT,
-    SPECIES_CHANDELURE,
-    SPECIES_AXEW,
-    SPECIES_FRAXURE,
-    SPECIES_HAXORUS,
-    SPECIES_CUBCHOO,
-    SPECIES_BEARTIC,
-    SPECIES_CRYOGONAL,
-    SPECIES_SHELMET,
-    SPECIES_ACCELGOR,
-    SPECIES_STUNFISK,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_STUNFISK_GALARIAN,
-#endif
-    SPECIES_MIENFOO,
-    SPECIES_MIENSHAO,
-    SPECIES_DRUDDIGON,
-    SPECIES_GOLETT,
-    SPECIES_GOLURK,
-    SPECIES_PAWNIARD,
-    SPECIES_BISHARP,
-    SPECIES_BOUFFALANT,
-    SPECIES_RUFFLET,
-    SPECIES_BRAVIARY,
-#ifdef PIT_GEN_9_MODE
-    SPECIES_BRAVIARY_HISUIAN,
-#endif
-    SPECIES_VULLABY,
-    SPECIES_MANDIBUZZ,
-    SPECIES_HEATMOR,
-    SPECIES_DURANT,
-    SPECIES_DEINO,
-    SPECIES_ZWEILOUS,
-    SPECIES_HYDREIGON,
-    SPECIES_LARVESTA,
-    SPECIES_VOLCARONA,
-    SPECIES_COBALION,
-    SPECIES_TERRAKION,
-    SPECIES_VIRIZION,
-    SPECIES_TORNADUS_INCARNATE,
-    SPECIES_THUNDURUS_INCARNATE,
-    SPECIES_RESHIRAM,
-    SPECIES_ZEKROM,
-    SPECIES_LANDORUS_INCARNATE,
-    SPECIES_KYUREM,
-    //SPECIES_KYUREM_WHITE,
-    //SPECIES_KYUREM_BLACK,
-    SPECIES_KELDEO_ORDINARY,
-    SPECIES_MELOETTA_ARIA,
-    SPECIES_GENESECT,
-#endif
-#if GEN_LATEST > 5
-    SPECIES_CHESPIN,
-    SPECIES_QUILLADIN,
-    SPECIES_CHESNAUGHT,
-    SPECIES_FENNEKIN,
-    SPECIES_BRAIXEN,
-    SPECIES_DELPHOX,
-    SPECIES_FROAKIE,
-    SPECIES_FROGADIER,
-    SPECIES_GRENINJA,
-    SPECIES_BUNNELBY,
-    SPECIES_DIGGERSBY,
-    SPECIES_FLETCHLING,
-    SPECIES_FLETCHINDER,
-    SPECIES_TALONFLAME,
-    SPECIES_SCATTERBUG_ICY_SNOW,
-    SPECIES_SPEWPA_ICY_SNOW,
-    SPECIES_VIVILLON_ICY_SNOW,
-    SPECIES_LITLEO,
-    SPECIES_PYROAR,
-    SPECIES_FLABEBE_RED_FLOWER,
-    SPECIES_FLOETTE_RED_FLOWER,
-    SPECIES_FLORGES_RED_FLOWER,
-    SPECIES_SKIDDO,
-    SPECIES_GOGOAT,
-    SPECIES_PANCHAM,
-    SPECIES_PANGORO,
-    SPECIES_FURFROU_NATURAL,
-    SPECIES_ESPURR,
-    SPECIES_MEOWSTIC_MALE,
-    SPECIES_HONEDGE,
-    SPECIES_DOUBLADE,
-    SPECIES_AEGISLASH_SHIELD,
-    SPECIES_SPRITZEE,
-    SPECIES_AROMATISSE,
-    SPECIES_SWIRLIX,
-    SPECIES_SLURPUFF,
-    SPECIES_INKAY,
-    SPECIES_MALAMAR,
-    SPECIES_BINACLE,
-    SPECIES_BARBARACLE,
-    SPECIES_SKRELP,
-    SPECIES_DRAGALGE,
-    SPECIES_CLAUNCHER,
-    SPECIES_CLAWITZER,
-    SPECIES_HELIOPTILE,
-    SPECIES_HELIOLISK,
-    SPECIES_TYRUNT,
-    SPECIES_TYRANTRUM,
-    SPECIES_AMAURA,
-    SPECIES_AURORUS,
-    SPECIES_SYLVEON,
-    SPECIES_HAWLUCHA,
-    SPECIES_DEDENNE,
-    SPECIES_CARBINK,
-    SPECIES_GOOMY,
-    SPECIES_SLIGGOO,
-    SPECIES_SLIGGOO_HISUIAN,
-    SPECIES_GOODRA,
-    SPECIES_GOODRA_HISUIAN,
-    SPECIES_KLEFKI,
-    SPECIES_PHANTUMP,
-    SPECIES_TREVENANT,
-    SPECIES_PUMPKABOO_AVERAGE,
-    SPECIES_GOURGEIST_AVERAGE,
-    SPECIES_BERGMITE,
-    SPECIES_AVALUGG,
-    SPECIES_AVALUGG_HISUIAN,
-    SPECIES_NOIBAT,
-    SPECIES_NOIVERN,
-    SPECIES_XERNEAS_NEUTRAL,
-    SPECIES_YVELTAL,
-    SPECIES_ZYGARDE_50_AURA_BREAK,
-    SPECIES_DIANCIE,
-    SPECIES_HOOPA_CONFINED,
-    SPECIES_VOLCANION,
-    SPECIES_ROWLET,
-    SPECIES_DARTRIX,
-    SPECIES_DECIDUEYE,
-    SPECIES_DECIDUEYE_HISUIAN,
-    SPECIES_LITTEN,
-    SPECIES_TORRACAT,
-    SPECIES_INCINEROAR,
-    SPECIES_POPPLIO,
-    SPECIES_BRIONNE,
-    SPECIES_PRIMARINA,
-    SPECIES_PIKIPEK,
-    SPECIES_TRUMBEAK,
-    SPECIES_TOUCANNON,
-    SPECIES_YUNGOOS,
-    SPECIES_GUMSHOOS,
-    SPECIES_GRUBBIN,
-    SPECIES_CHARJABUG,
-    SPECIES_VIKAVOLT,
-    SPECIES_CRABRAWLER,
-    SPECIES_CRABOMINABLE,
-    SPECIES_ORICORIO_BAILE,
-    SPECIES_CUTIEFLY,
-    SPECIES_RIBOMBEE,
-    SPECIES_ROCKRUFF,
-    SPECIES_LYCANROC_MIDDAY,
-    SPECIES_WISHIWASHI_SOLO,
-    SPECIES_MAREANIE,
-    SPECIES_TOXAPEX,
-    SPECIES_MUDBRAY,
-    SPECIES_MUDSDALE,
-    SPECIES_DEWPIDER,
-    SPECIES_ARAQUANID,
-    SPECIES_FOMANTIS,
-    SPECIES_LURANTIS,
-    SPECIES_MORELULL,
-    SPECIES_SHIINOTIC,
-    SPECIES_SALANDIT,
-    SPECIES_SALAZZLE,
-    SPECIES_STUFFUL,
-    SPECIES_BEWEAR,
-    SPECIES_BOUNSWEET,
-    SPECIES_STEENEE,
-    SPECIES_TSAREENA,
-    SPECIES_COMFEY,
-    SPECIES_ORANGURU,
-    SPECIES_PASSIMIAN,
-    SPECIES_WIMPOD,
-    SPECIES_GOLISOPOD,
-    SPECIES_SANDYGAST,
-    SPECIES_PALOSSAND,
-    SPECIES_PYUKUMUKU,
-    SPECIES_TYPE_NULL,
-    SPECIES_SILVALLY_NORMAL,
-    SPECIES_MINIOR_METEOR_RED,
-    SPECIES_KOMALA,
-    SPECIES_TURTONATOR,
-    SPECIES_TOGEDEMARU,
-    SPECIES_MIMIKYU_DISGUISED,
-    SPECIES_BRUXISH,
-    SPECIES_DRAMPA,
-    SPECIES_DHELMISE,
-    SPECIES_JANGMO_O,
-    SPECIES_HAKAMO_O,
-    SPECIES_KOMMO_O,
-    SPECIES_TAPU_KOKO,
-    SPECIES_TAPU_LELE,
-    SPECIES_TAPU_BULU,
-    SPECIES_TAPU_FINI,
-    SPECIES_COSMOG,
-    SPECIES_COSMOEM,
-    SPECIES_SOLGALEO,
-    SPECIES_LUNALA,
-    SPECIES_NIHILEGO,
-    SPECIES_BUZZWOLE,
-    SPECIES_PHEROMOSA,
-    SPECIES_XURKITREE,
-    SPECIES_CELESTEELA,
-    SPECIES_KARTANA,
-    SPECIES_GUZZLORD,
-    SPECIES_NECROZMA,
-    SPECIES_MAGEARNA,
-    SPECIES_MARSHADOW,
-    SPECIES_POIPOLE,
-    SPECIES_NAGANADEL,
-    SPECIES_STAKATAKA,
-    SPECIES_BLACEPHALON,
-    SPECIES_ZERAORA,
-    SPECIES_MELTAN,
-    SPECIES_MELMETAL,
-    SPECIES_GROOKEY,
-    SPECIES_THWACKEY,
-    SPECIES_RILLABOOM,
-    SPECIES_SCORBUNNY,
-    SPECIES_RABOOT,
-    SPECIES_CINDERACE,
-    SPECIES_SOBBLE,
-    SPECIES_DRIZZILE,
-    SPECIES_INTELEON,
-    SPECIES_SKWOVET,
-    SPECIES_GREEDENT,
-    SPECIES_ROOKIDEE,
-    SPECIES_CORVISQUIRE,
-    SPECIES_CORVIKNIGHT,
-    SPECIES_BLIPBUG,
-    SPECIES_DOTTLER,
-    SPECIES_ORBEETLE,
-    SPECIES_NICKIT,
-    SPECIES_THIEVUL,
-    SPECIES_GOSSIFLEUR,
-    SPECIES_ELDEGOSS,
-    SPECIES_WOOLOO,
-    SPECIES_DUBWOOL,
-    SPECIES_CHEWTLE,
-    SPECIES_DREDNAW,
-    SPECIES_YAMPER,
-    SPECIES_BOLTUND,
-    SPECIES_ROLYCOLY,
-    SPECIES_CARKOL,
-    SPECIES_COALOSSAL,
-    SPECIES_APPLIN,
-    SPECIES_FLAPPLE,
-    SPECIES_APPLETUN,
-    SPECIES_SILICOBRA,
-    SPECIES_SANDACONDA,
-    SPECIES_CRAMORANT,
-    SPECIES_ARROKUDA,
-    SPECIES_BARRASKEWDA,
-    SPECIES_TOXEL,
-    SPECIES_TOXTRICITY_AMPED,
-    SPECIES_SIZZLIPEDE,
-    SPECIES_CENTISKORCH,
-    SPECIES_CLOBBOPUS,
-    SPECIES_GRAPPLOCT,
-    SPECIES_SINISTEA_PHONY,
-    SPECIES_POLTEAGEIST_PHONY,
-    SPECIES_HATENNA,
-    SPECIES_HATTREM,
-    SPECIES_HATTERENE,
-    SPECIES_IMPIDIMP,
-    SPECIES_MORGREM,
-    SPECIES_GRIMMSNARL,
-    SPECIES_OBSTAGOON,
-    SPECIES_PERRSERKER,
-    SPECIES_CURSOLA,
-    SPECIES_SIRFETCHD,
-    SPECIES_MR_RIME,
-    SPECIES_RUNERIGUS,
-    SPECIES_MILCERY,
-    SPECIES_ALCREMIE_STRAWBERRY_VANILLA_CREAM,
-    SPECIES_FALINKS,
-    SPECIES_PINCURCHIN,
-    SPECIES_SNOM,
-    SPECIES_FROSMOTH,
-    SPECIES_STONJOURNER,
-    SPECIES_EISCUE_ICE_FACE,
-    SPECIES_INDEEDEE_MALE,
-    SPECIES_MORPEKO_FULL_BELLY,
-    SPECIES_CUFANT,
-    SPECIES_COPPERAJAH,
-    SPECIES_DRACOZOLT,
-    SPECIES_ARCTOZOLT,
-    SPECIES_DRACOVISH,
-    SPECIES_ARCTOVISH,
-    SPECIES_DURALUDON,
-    SPECIES_DREEPY,
-    SPECIES_DRAKLOAK,
-    SPECIES_DRAGAPULT,
-    SPECIES_ZACIAN_HERO_OF_MANY_BATTLES,
-    SPECIES_ZAMAZENTA_HERO_OF_MANY_BATTLES,
-    SPECIES_ETERNATUS,
-    SPECIES_KUBFU,
-    SPECIES_URSHIFU_SINGLE_STRIKE_STYLE,
-    SPECIES_ZARUDE,
-    SPECIES_REGIELEKI,
-    SPECIES_REGIDRAGO,
-    SPECIES_GLASTRIER,
-    SPECIES_SPECTRIER,
-    SPECIES_CALYREX,
-    SPECIES_WYRDEER,
-    SPECIES_KLEAVOR,
-    SPECIES_URSALUNA,
-    SPECIES_URSALUNA_BLOODMOON,
-    SPECIES_BASCULEGION_MALE,
-    SPECIES_SNEASLER,
-    SPECIES_OVERQWIL,
-    SPECIES_ENAMORUS_INCARNATE,
-    SPECIES_GRENINJA_BATTLE_BOND,
-    SPECIES_ZYGARDE_10_AURA_BREAK,
-    SPECIES_ROCKRUFF_OWN_TEMPO,
-    SPECIES_LYCANROC_MIDNIGHT,
-    SPECIES_LYCANROC_DUSK,
-    SPECIES_SPRIGATITO,
-    SPECIES_FLORAGATO,
-    SPECIES_MEOWSCARADA,
-    SPECIES_FUECOCO,
-    SPECIES_CROCALOR,
-    SPECIES_SKELEDIRGE,
-    SPECIES_QUAXLY,
-    SPECIES_QUAXWELL,
-    SPECIES_QUAQUAVAL,
-    SPECIES_LECHONK,
-    SPECIES_OINKOLOGNE_MALE,
-    SPECIES_TAROUNTULA,
-    SPECIES_SPIDOPS,
-    SPECIES_NYMBLE,
-    SPECIES_LOKIX,
-    SPECIES_PAWMI,
-    SPECIES_PAWMO,
-    SPECIES_PAWMOT,
-    SPECIES_TANDEMAUS,
-    SPECIES_MAUSHOLD_FAMILY_OF_THREE,
-    SPECIES_MAUSHOLD_FAMILY_OF_FOUR,
-    SPECIES_FIDOUGH,
-    SPECIES_DACHSBUN,
-    SPECIES_SMOLIV,
-    SPECIES_DOLLIV,
-    SPECIES_ARBOLIVA,
-    SPECIES_SQUAWKABILLY_GREEN_PLUMAGE,
-    SPECIES_NACLI,
-    SPECIES_NACLSTACK,
-    SPECIES_GARGANACL,
-    SPECIES_CHARCADET,
-    SPECIES_ARMAROUGE,
-    SPECIES_CERULEDGE,
-    SPECIES_TADBULB,
-    SPECIES_BELLIBOLT,
-    SPECIES_WATTREL,
-    SPECIES_KILOWATTREL,
-    SPECIES_MASCHIFF,
-    SPECIES_MABOSSTIFF,
-    SPECIES_SHROODLE,
-    SPECIES_GRAFAIAI,
-    SPECIES_BRAMBLIN,
-    SPECIES_BRAMBLEGHAST,
-    SPECIES_TOEDSCOOL,
-    SPECIES_TOEDSCRUEL,
-    SPECIES_KLAWF,
-    SPECIES_CAPSAKID,
-    SPECIES_SCOVILLAIN,
-    SPECIES_RELLOR,
-    SPECIES_RABSCA,
-    SPECIES_FLITTLE,
-    SPECIES_ESPATHRA,
-    SPECIES_TINKATINK,
-    SPECIES_TINKATUFF,
-    SPECIES_TINKATON,
-    SPECIES_WIGLETT,
-    SPECIES_WUGTRIO,
-    SPECIES_BOMBIRDIER,
-    SPECIES_FINIZEN,
-    SPECIES_PALAFIN_ZERO,
-    SPECIES_PALAFIN_HERO,
-    SPECIES_VAROOM,
-    SPECIES_REVAVROOM,
-    SPECIES_CYCLIZAR,
-    SPECIES_ORTHWORM,
-    SPECIES_GLIMMET,
-    SPECIES_GLIMMORA,
-    SPECIES_GREAVARD,
-    SPECIES_HOUNDSTONE,
-    SPECIES_FLAMIGO,
-    SPECIES_CETODDLE,
-    SPECIES_CETITAN,
-    SPECIES_VELUZA,
-    SPECIES_DONDOZO,
-    SPECIES_TATSUGIRI_CURLY,
-    SPECIES_TATSUGIRI_DROOPY,
-    SPECIES_TATSUGIRI_STRETCHY,
-    SPECIES_ANNIHILAPE,
-    SPECIES_CLODSIRE,
-    SPECIES_FARIGIRAF,
-    SPECIES_DUDUNSPARCE_TWO_SEGMENT,
-    SPECIES_DUDUNSPARCE_THREE_SEGMENT,
-    SPECIES_KINGAMBIT,
-    SPECIES_GREAT_TUSK,
-    SPECIES_SCREAM_TAIL,
-    SPECIES_BRUTE_BONNET,
-    SPECIES_FLUTTER_MANE,
-    SPECIES_SLITHER_WING,
-    SPECIES_SANDY_SHOCKS,
-    SPECIES_IRON_TREADS,
-    SPECIES_IRON_BUNDLE,
-    SPECIES_IRON_HANDS,
-    SPECIES_IRON_JUGULIS,
-    SPECIES_IRON_MOTH,
-    SPECIES_IRON_THORNS,
-    SPECIES_FRIGIBAX,
-    SPECIES_ARCTIBAX,
-    SPECIES_BAXCALIBUR,
-    SPECIES_GIMMIGHOUL_CHEST,
-    SPECIES_GHOLDENGO,
-    SPECIES_WO_CHIEN,
-    SPECIES_CHIEN_PAO,
-    SPECIES_TING_LU,
-    SPECIES_CHI_YU,
-    SPECIES_ROARING_MOON,
-    SPECIES_IRON_VALIANT,
-    SPECIES_KORAIDON,
-    SPECIES_MIRAIDON,
-    SPECIES_WALKING_WAKE,
-    SPECIES_IRON_LEAVES,
-    SPECIES_DIPPLIN,
-    SPECIES_POLTCHAGEIST_COUNTERFEIT,
-    SPECIES_SINISTCHA_UNREMARKABLE,
-    SPECIES_OKIDOGI,
-    SPECIES_MUNKIDORI,
-    SPECIES_FEZANDIPITI,
-    SPECIES_OGERPON_TEAL_MASK,
-    SPECIES_ARCHALUDON,
-    SPECIES_HYDRAPPLE,
-    SPECIES_GOUGING_FIRE,
-    SPECIES_RAGING_BOLT,
-    SPECIES_IRON_BOULDER,
-    SPECIES_IRON_CROWN,
-    SPECIES_TERAPAGOS_NORMAL,
-    SPECIES_PECHARUNT
-    #endif
-    // SPECIES_EGG       ,
-};
-
-u16 AccessValidSpeciesArrayIndex(u16 index)
-{
-    return sRandomSpeciesLegendary[index % RANDOM_SPECIES_COUNT_LEGENDARY];
-}
-
-u16 GetIndexOfSpeciesInValidSpeciesArray(u16 species)
-{   
-    for(int i = 0; i < RANDOM_SPECIES_COUNT_LEGENDARY; i++)
-    {
-        if(sRandomSpeciesLegendary[i] == species)
-        {
-            return i;
-        }
-    }
-    return 0xFFFF;
-}
-
-u16 GetSpeciesRandomSeeded(u16 species)
-{
-    return sRandomSpeciesLegendary[RandomSeededModulo2(species, RANDOM_SPECIES_COUNT_LEGENDARY)];
-}
-
-u16 GetSpeciesRandomNotSeeded(u16 species)
-{
-    return sRandomSpeciesLegendary[RandomModulo(species, RANDOM_SPECIES_COUNT_LEGENDARY)];
-}
-
-u16 GetRandomMove(u16 move, u16 species)
-{
-    u16 val = RandomSeededModulo2(move + species, GetRandomValidMovesCount());
-    u16 final = gRandomValidMoves[val];
-    return final;
-}
-
-u16 GetRandomMoveNotSeeded(u16 move, u16 species)
-{
-    u16 val = RandomSeededModulo2((Random() % GetRandomValidMovesCount()) + move + species, GetRandomValidMovesCount());
-    u16 final = gRandomValidMoves[val];
-    return final;
-}
-
-const u8 *GetMoveName(u16 moveId)
-{
-    return gMovesInfo[moveId].name;
 }
 
 bool8 IsSpeciesLegendary(u16 species)
@@ -2621,14 +1351,31 @@ bool8 IsSpeciesParadoxMon(u16 species)
     return FALSE;
 }
 
-u32 GetMaxNumberOfSpeciesAvatars(void)
+u16 AccessValidSpeciesArrayIndex(u16 index)
 {
-    return AVATAR_SPECIES_COUNT;
+    return GetSpeciesFromRandomArray(index % GetMaxNumberOfSpecies());
 }
 
-u32 GetMaxNumberOfSpecies(void)
+u16 GetIndexOfSpeciesInValidSpeciesArray(u16 species)
+{   
+    for(int i = 0; i < GetMaxNumberOfSpecies(); i++)
+    {
+        if(GetSpeciesFromRandomArray(i) == species)
+        {
+            return i;
+        }
+    }
+    return 0xFFFF;
+}
+
+u16 GetSpeciesRandomSeeded(u16 species)
 {
-    return RANDOM_SPECIES_COUNT_LEGENDARY;
+    return GetSpeciesFromRandomArray(RandomSeededModulo2(species, GetMaxNumberOfSpecies()));
+}
+
+u16 GetSpeciesRandomNotSeeded(u16 species)
+{
+    return GetSpeciesFromRandomArray(RandomModulo(species, GetMaxNumberOfSpecies()));
 }
 
 u8 GetCurrentMaxMonGeneratedCount(void)
@@ -3997,34 +2744,26 @@ const u16 gRandomValidMoves[] =
     #endif
 };
 
+//**********************
 
-
-//
-//		Random Battle Weathers
-//
-
-//tx_randomizer_and_challenges
-#define RANDOM_BATTLE_WEATHER_COUNT ARRAY_COUNT(sRandomBattleWeathers)
-static const u16 sRandomBattleWeathers[] =
+u16 GetRandomMove(u16 move, u16 species)
 {
-    WEATHER_NONE,
-    WEATHER_RAIN,
-    WEATHER_SANDSTORM,
-    WEATHER_DROUGHT,
-    WEATHER_SNOW,
-};
-
-u16 GetRandomBattleWeather(void)
-{
-    u16 battleWeather;
-
-    battleWeather = (Random() % (RANDOM_BATTLE_WEATHER_COUNT));
-    battleWeather = sRandomBattleWeathers[battleWeather];
-    return battleWeather;
+    u16 val = RandomSeededModulo2(move + species, GetRandomValidMovesCount());
+    u16 final = gRandomValidMoves[val];
+    return final;
 }
 
+u16 GetRandomMoveNotSeeded(u16 move, u16 species)
+{
+    u16 val = RandomSeededModulo2((Random() % GetRandomValidMovesCount()) + move + species, GetRandomValidMovesCount());
+    u16 final = gRandomValidMoves[val];
+    return final;
+}
 
-//**********************
+const u8 *GetMoveName(u16 moveId)
+{
+    return gMovesInfo[moveId].name;
+}
 
 u16 GetRandomValidMovesCount(void)
 {
@@ -4532,3 +3271,28 @@ u16 GetRandomSecretBaseID(void)
     }
     return secretBaseId;
 }
+
+//
+//		Random Battle Weathers
+//
+
+//tx_randomizer_and_challenges
+#define RANDOM_BATTLE_WEATHER_COUNT ARRAY_COUNT(sRandomBattleWeathers)
+static const u16 sRandomBattleWeathers[] =
+{
+    WEATHER_NONE,
+    WEATHER_RAIN,
+    WEATHER_SANDSTORM,
+    WEATHER_DROUGHT,
+    WEATHER_SNOW,
+};
+
+u16 GetRandomBattleWeather(void)
+{
+    u16 battleWeather;
+
+    battleWeather = (Random() % (RANDOM_BATTLE_WEATHER_COUNT));
+    battleWeather = sRandomBattleWeathers[battleWeather];
+    return battleWeather;
+}
+
